requirements.txt
numpy<2
flask
gunicorn
pandas
scikit-learn
scikit-surprise